Build: unify build.sh and buildn.sh
build.sh now checks the OS and supports Bash on Windows as well.
buildn.sh is therefore no longer needed.
See also fsprojects/ProjectScaffold#76

+++ build.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
-if [ ! -f packages/FAKE/tools/FAKE.exe ]; then
-  mono --runtime=v4.0 tools/NuGet/NuGet.exe install FAKE -OutputDirectory packages -ExcludeVersion
+if test "$OS" = "Windows_NT"
+then
+  # use .Net
+  [ ! -f packages/FAKE/tools/FAKE.exe ] && tools/NuGet/NuGet.exe install FAKE -OutputDirectory packages -ExcludeVersion
+  packages/FAKE/tools/FAKE.exe build.fsx $@
+else
+  # use mono
+  [ ! -f packages/FAKE/tools/FAKE.exe ] && mono --runtime=v4.0 tools/NuGet/NuGet.exe install FAKE -OutputDirectory packages -ExcludeVersion
+  mono --runtime=v4.0 packages/FAKE/tools/FAKE.exe $@ --fsiargs -d:MONO build.fsx
 fi
-mono --runtime=v4.0 packages/FAKE/tools/FAKE.exe build.fsx $@
